sd: clean up sizes of some SD data fields
Clean up sizes of some SD data fields, as the max value of these fields
is limited by the SD specification

Specifically, the limits are as follows:
num_io: 0-7, 3 bits (SDIO only)
relative_addr: 16 bits (SDMMC/MMC)
block_size: 12 bits (Max of 2KB, uint16_t used)
sd_version: 8 bits (currently at version 3)
card_speed: 8 bits (could potentially be reduced in size)

+++ subsys/sd/sdmmc.c
@@ -19,7 +19,7 @@
 LOG_MODULE_DECLARE(sd, CONFIG_SD_LOG_LEVEL);
 
 static inline void sdmmc_decode_scr(struct sd_scr *scr,
-	uint32_t *raw_scr, uint32_t *version)
+	uint32_t *raw_scr, uint8_t *version)
 {
 	uint32_t tmp_version = 0;
